Provide support for the last five go versions via travis
Looking the gorilla/websocket travis settings on https://github.com/gorilla/websocket/blob/master/.travis.yml, we can see that the project supports go 1.7+ and that's could be the reason why our project is breaking on go 1.6 .
In this commit I'm setting four recent go versions (1.7 - 1.10) and the latest one to ensure that project is running correctly in those versions. As you can see it removes the 1.6 support and maybe we need to communicate that on readme.