Use an empty struct channel instead of Mutex to protect the Observer.disposed.
This changes o.disposed in Observer and SingleObserver from a Mutex
protected bool to a channel of struct{}. Once the channel is closed,
that's considered "disposed", and avoids reglar mutex synchronisation.

Included is a benchmark (go run -race -bench=. ./...) to test what I
assume is a regular use of IsDisposed on Observer and SingleObserver.

The original code:
BenchmarkIsDisposed-8             300000              4776 ns/op

After this change:
BenchmarkIsDisposed-8            1000000              1138 ns/op

It's roughly 4x faster using a channel instead of a Mutex.